Move dat core into its own daemon thread (WIP)
