+++ dat/daemon/extensions.js
