+++ dat/daemon/folder-sync.js
@@ -1,4 +1,3 @@
-const globals = require('../globals')
 const bytes = require('bytes')
 const dft = require('diff-file-tree')
 const diff = require('diff')
@@ -9,10 +8,10 @@ const EventEmitter = require('events')
 const pda = require('pauls-dat-api')
 const mkdirp = require('mkdirp')
 const {toAnymatchRules} = require('@beaker/datignore')
-const settingsDb = require('../dbs/settings')
-const {isFileNameBinary, isFileContentBinary} = require('../lib/mime')
-const lock = require('../lib/lock')
-const scopedFSes = require('../lib/scoped-fses')
+// const settingsDb = require('../dbs/settings') DAEMON
+const {isFileNameBinary, isFileContentBinary} = require('../../lib/mime')
+const lock = require('../../lib/lock')
+const scopedFSes = require('../../lib/scoped-fses')
 const {
   NotFoundError,
   NotAFolderError,
@@ -279,12 +278,13 @@ exports.diffFile = async function (archive, filepath) {
 
 // validate a path to be used for sync
 exports.assertSafePath = async function (p) {
+  // DAEMON
   // check whether this is an OS path
-  for (let disallowedSavePath of globals.disallowedSavePaths) {
-    if (path.normalize(p) === path.normalize(disallowedSavePath)) {
-      throw new ProtectedFileNotWritableError(`This is a protected folder. Please pick another folder or subfolder.`)
-    }
-  }
+  // for (let disallowedSavePath of globals.disallowedSavePaths) {
+  //   if (path.normalize(p) === path.normalize(disallowedSavePath)) {
+  //     throw new ProtectedFileNotWritableError(`This is a protected folder. Please pick another folder or subfolder.`)
+  //   }
+  // }
 
   // stat the folder
   const stat = await new Promise(resolve => {
@@ -304,8 +304,9 @@ exports.assertSafePath = async function (p) {
 const readDatIgnore = exports.readDatIgnore = async function (fs) {
   var rulesRaw = await readFile(fs, '.datignore')
   if (!rulesRaw) {
+    // DAEMON
     // TODO remove this? we're supposed to only use .datignore but many archives wont have one at first -prf
-    rulesRaw = await settingsDb.get('default_dat_ignore')
+    // rulesRaw = await settingsDb.get('default_dat_ignore')
   }
   return toAnymatchRules(rulesRaw)
 }
