+++ dat/daemon/logging-utils.js
