+++ dat/daemon/manifest.js
@@ -1,0 +1,11 @@
+module.exports = {
+  configureArchive: 'promise',
+  setBandwidthThrottle: 'promise',
+  getDebugLog: 'promise',
+  loadArchive: 'promise',
+  unloadArchive: 'promise',
+  callArchiveAsyncMethod: 'async',
+  callArchiveReadStreamMethod: 'readable',
+  callArchiveWriteStreamMethod: 'writable',
+  clearFileCache: 'promise'
+}
