+++ dat/index.js
@@ -1,7 +1,6 @@
 module.exports = {
   debug: require('./debugging'),
   dns: require('./dns'),
-  folderSync: require('./folder-sync'),
   garbageCollector: require('./garbage-collector'),
   library: require('./library'),
   protocol: require('./protocol'),
