+++ dat/library.js
@@ -1,4 +1,3 @@
-const crypto = require('crypto')
 const emitStream = require('emit-stream')
 const EventEmitter = require('events')
 const datEncoding = require('dat-encoding')
@@ -6,37 +5,18 @@ const pify = require('pify')
 const pda = require('pauls-dat-api')
 const signatures = require('sodium-signatures')
 const parseDatURL = require('parse-dat-url')
-const through = require('through2')
-const split = require('split2')
-const concat = require('concat-stream')
-const CircularAppendFile = require('circular-append-file')
-const debug = require('../lib/debug-logger').debugLogger('dat')
-const throttle = require('lodash.throttle')
 const debounce = require('lodash.debounce')
-const isEqual = require('lodash.isequal')
-const pump = require('pump')
+
+// dbs
 const siteData = require('../dbs/sitedata')
 const settingsDb = require('../dbs/settings')
+const archivesDb = require('../dbs/archives')
 
 // dat modules
-const archivesDb = require('../dbs/archives')
-const datStorage = require('./storage')
 const datGC = require('./garbage-collector')
-const folderSync = require('./folder-sync')
-const {addArchiveSwarmLogging} = require('./logging-utils')
-const datExtensions = require('./extensions')
-const hypercoreProtocol = require('hypercore-protocol')
-const hyperdrive = require('hyperdrive')
-
-// network modules
-const swarmDefaults = require('datland-swarm-defaults')
-const discoverySwarm = require('discovery-swarm')
-const networkSpeed = require('hyperdrive-network-speed')
-const {ThrottleGroup} = require('stream-throttle')
 
 // file modules
 const mkdirp = require('mkdirp')
-const jetpack = require('fs-jetpack')
 const scopedFSes = require('../lib/scoped-fses')
 
 // constants
@@ -44,32 +24,25 @@ const scopedFSes = require('../lib/scoped-fses')
 
 const {
   DAT_HASH_REGEX,
-  DAT_SWARM_PORT,
   DAT_PRESERVED_FIELDS_ON_FORK
 } = require('../lib/const')
 const {InvalidURLError} = require('beaker-error-constants')
+const DAT_DAEMON_MANIFEST = require('./daemon/manifest')
 
 // globals
 // =
 
-var networkId = crypto.randomBytes(32)
 var archives = {} // in-memory cache of archive objects. key -> archive
-var archivesByDKey = {} // same, but discoveryKey -> archive
 var archiveLoadPromises = {} // key -> promise
 var archivesEvents = new EventEmitter()
-var debugEvents = new EventEmitter()
-var debugLogFile
-var archiveSwarm
-
-var upThrottleGroup
-var downThrottleGroup
+var daemon
 
 // exported API
 // =
 
-exports.setup = async function setup ({logfilePath}) {
-  await datStorage.setup()
-  debugLogFile = CircularAppendFile(logfilePath, {maxSize: 1024 /* 1kb */ * 1024 /* 1mb */ * 50 /* 50mb */ })
+exports.setup = async function setup ({rpcAPI, datDaemonWc}) {
+  // connect to the daemon
+  daemon = rpcAPI.importAPI('dat-daemon', DAT_DAEMON_MANIFEST, {wc: datDaemonWc})
 
   // wire up event handlers
   archivesDb.on('update:archive-user-settings', async (key, userSettings, newUserSettings) => {
@@ -95,59 +68,40 @@ exports.setup = async function setup ({logfilePath}) {
     }
 
     // update the download based on these settings
-    var archive = getArchive(key)
-    if (archive) {
-      configureNetwork(archive, userSettings)
-      configureAutoDownload(archive, userSettings)
-      configureLocalSync(archive, userSettings)
-    }
+    daemon.configureArchive(key, userSettings)
   })
-  folderSync.events.on('sync', (key, direction) => {
-    archivesEvents.emit('folder-synced', {
-      details: {
-        url: `dat://${datEncoding.toStr(key)}`,
-        direction
-      }
-    })
-  })
-  folderSync.events.on('error', (key, err) => {
-    archivesEvents.emit('folder-sync-error', {
-      details: {
-        url: `dat://${datEncoding.toStr(key)}`,
-        name: err.name,
-        message: err.message
-      }
-    })
-  })
+
+  // DAEMON
+  // folderSync.events.on('sync', (key, direction) => {
+  //   archivesEvents.emit('folder-synced', {
+  //     details: {
+  //       url: `dat://${datEncoding.toStr(key)}`,
+  //       direction
+  //     }
+  //   })
+  // })
+  // folderSync.events.on('error', (key, err) => {
+  //   archivesEvents.emit('folder-sync-error', {
+  //     details: {
+  //       url: `dat://${datEncoding.toStr(key)}`,
+  //       name: err.name,
+  //       message: err.message
+  //     }
+  //   })
+  // })
 
   // configure the bandwidth throttle
   settingsDb.getAll().then(({dat_bandwidth_limit_up, dat_bandwidth_limit_down}) => {
-    setBandwidthThrottle({
+    daemon.setBandwidthThrottle({
       up: dat_bandwidth_limit_up,
       down: dat_bandwidth_limit_down
     })
   })
-  settingsDb.on('set:dat_bandwidth_limit_up', up => setBandwidthThrottle({up}))
-  settingsDb.on('set:dat_bandwidth_limit_down', down => setBandwidthThrottle({down}))
+  settingsDb.on('set:dat_bandwidth_limit_up', up => daemon.setBandwidthThrottle({up}))
+  settingsDb.on('set:dat_bandwidth_limit_down', down => daemon.setBandwidthThrottle({down}))
 
-  // setup extension messages
-  datExtensions.setup()
-
-  // setup the archive swarm
+  // start the GC manager
   datGC.setup()
-  archiveSwarm = discoverySwarm(swarmDefaults({
-    id: networkId,
-    hash: false,
-    utp: true,
-    tcp: true,
-    dht: false,
-    connect: connectReplicationStream,
-    stream: createReplicationStream
-  }))
-  addArchiveSwarmLogging({archivesByDKey, log, archiveSwarm})
-  archiveSwarm.once('error', () => archiveSwarm.listen(0))
-  archiveSwarm.listen(DAT_SWARM_PORT)
-  archiveSwarm.on('error', error => log(null, {event: 'swarm-error', message: error.toString()}))
 }
 
 exports.loadSavedArchives = function () {
@@ -171,40 +125,17 @@ exports.loadSavedArchives = function () {
   )
 }
 
-// up/down are in MB/s
-const setBandwidthThrottle = exports.setBandwidthThrottle = function ({up, down}) {
-  if (typeof up !== 'undefined') {
-    debug(`Throttling upload to ${up} MB/s`)
-    upThrottleGroup = up ? new ThrottleGroup({rate: up * 1e6}) : null
-  }
-  if (typeof down !== 'undefined') {
-    debug(`Throttling download to ${down} MB/s`)
-    downThrottleGroup = down ? new ThrottleGroup({rate: down * 1e6}) : null
-  }
-}
-
 exports.createEventStream = function createEventStream () {
   return emitStream(archivesEvents)
 }
 
 exports.getDebugLog = function getDebugLog (key) {
-  return new Promise((resolve, reject) => {
-    let rs = debugLogFile.createReadStream()
-    rs
-      .pipe(split())
-      .pipe(through({encoding: 'utf8', decodeStrings: false}, (data, _, cb) => {
-        if (data && (!key || data.startsWith(key))) {
-          return cb(null, data.slice(64) + '\n')
-        }
-        cb()
-      }))
-      .pipe(concat({encoding: 'string'}, resolve))
-    rs.on('error', reject)
-  })
+  return daemon.getDebugLog(key)
 }
 
 exports.createDebugStream = function createDebugStream () {
-  return emitStream(debugEvents)
+  // DAEMON
+  // return emitStream(debugEvents)
 }
 
 // read metadata for the archive, and store it in the meta db
@@ -219,17 +150,16 @@ const pullLatestArchiveMeta = exports.pullLatestArchiveMeta = async function pul
     var [manifest, oldMeta] = await Promise.all([
       pda.readManifest(archive).catch(_ => {}),
       archivesDb.getMeta(key),
-      updateSizeTracking(archive)
+      // daemon.updateSizeTracking(archive) DAEMON
     ])
     manifest = archive.manifest = manifest || {}
     var {title, description, type} = manifest
-    var isOwner = archive.writable
-    var size = archive.size || 0
+    var isOwner = false // archive.writable DAEMON
+    var size = archive.size || 0 // DAEMON
     var mtime = updateMTime ? Date.now() : oldMeta.mtime
 
     // write the record
     var details = {title, description, type, mtime, size, isOwner}
-    debug('Writing meta', details)
     await archivesDb.setMeta(key, details)
 
     // emit the updated event
@@ -381,85 +311,41 @@ async function loadArchiveInner (key, secretKey, userSettings = null) {
   var metaPath = archivesDb.getArchiveMetaPath(key)
   mkdirp.sync(metaPath)
 
-  // create the archive instance
-  var archive = hyperdrive(datStorage.create(metaPath), key, {
-    sparse: true,
+  // load the archive in the daemon
+  var archiveInfo = await daemon.loadArchive({
+    key,
     secretKey,
-    metadataStorageCacheSize: 0,
-    contentStorageCacheSize: 0,
-    treeCacheSize: 2048
+    metaPath,
+    userSettings
   })
-  archive.on('error', err => {
-    let k = key.toString('hex')
-    log(k, {event: 'archive-error', message: err.toString()})
-    console.error('Error in archive', k, err)
-    debug('Error in archive', k, err)
-  })
-  archive.metadata.on('peer-add', () => onNetworkChanged(archive))
-  archive.metadata.on('peer-remove', () => onNetworkChanged(archive))
-  archive.networkStats = networkSpeed(archive)
-  archive.replicationStreams = [] // list of all active replication streams
-  archive.peerHistory = [] // samples of the peer count
 
-  // wait for ready
-  await new Promise((resolve, reject) => {
-    archive.ready(err => {
-      if (err) reject(err)
-      else resolve()
-    })
-  })
-  await updateSizeTracking(archive)
+  // create the archive proxy instance
+  var archive = createArchiveProxy(key, archiveInfo)
+
+  // update db
   archivesDb.touch(key).catch(err => console.error('Failed to update lastAccessTime for archive', key, err))
-
-  // attach extensions
-  datExtensions.attach(archive)
-
-  // store in the discovery listing, so the swarmer can find it
-  // but not yet in the regular archives listing, because it's not fully loaded
-  archivesByDKey[datEncoding.toStr(archive.discoveryKey)] = archive
-
-  // setup the archive based on current settings
-  configureNetwork(archive, userSettings)
-  configureAutoDownload(archive, userSettings)
-  configureLocalSync(archive, userSettings)
-
-  // await initial metadata sync if not the owner
-  if (!archive.writable && !archive.metadata.length) {
-    // wait to receive a first update
-    await new Promise((resolve, reject) => {
-      archive.metadata.update(err => {
-        if (err) reject(err)
-        else resolve()
-      })
-    })
-  }
-  if (!archive.writable) {
-    // always download all metadata
-    archive.metadata.download({start: 0, end: -1})
-  }
-
-  // pull meta
   await pullLatestArchiveMeta(archive)
 
   // wire up events
   archive.pullLatestArchiveMeta = debounce(opts => pullLatestArchiveMeta(archive, opts), 1e3)
-  archive.fileActStream = pda.watch(archive)
-  archive.fileActStream.on('data', ([event, {path}]) => {
-    if (event === 'changed') {
-      archive.pullLatestArchiveMeta({updateMTime: true})
-      let syncSettings = archive.localSyncSettings
-      if (syncSettings) {
-        // need to sync this change to the local folder
-        if (syncSettings.autoPublish) {
-          // bidirectional sync: use the sync queue
-          folderSync.queueSyncEvent(archive, {toFolder: true})
-        } else {
-          // preview mode: just write this update to disk
-          folderSync.syncArchiveToFolder(archive, {paths: [path], shallow: false})
-        }
-      }
-    }
-  })
+  // DAEMON
+  // archive.fileActStream = pda.watch(archive)
+  // archive.fileActStream.on('data', ([event, {path}]) => {
+  //   if (event === 'changed') {
+  //     archive.pullLatestArchiveMeta({updateMTime: true})
+  //     let syncSettings = archive.localSyncSettings
+  //     if (syncSettings) {
+  //       // need to sync this change to the local folder
+  //       if (syncSettings.autoPublish) {
+  //         // bidirectional sync: use the sync queue
+  //         folderSync.queueSyncEvent(archive, {toFolder: true})
+  //       } else {
+  //         // preview mode: just write this update to disk
+  //         folderSync.syncArchiveToFolder(archive, {paths: [path], shallow: false})
+  //       }
+  //     }
+  //   }
+  // })
 
   // now store in main archives listing, as loaded
   archives[datEncoding.toStr(archive.key)] = archive
@@ -475,32 +361,33 @@ const getArchiveCheckout = exports.getArchiveCheckout = function getArchiveCheck
   var isHistoric = false
   var isPreview = false
   var checkoutFS = archive
-  if (version) {
-    let seq = parseInt(version)
-    if (Number.isNaN(seq)) {
-      if (version === 'latest') {
-        // ignore, we use latest by default
-      } else if (version === 'preview') {
-        if (archive.localSyncSettings) {
-          // checkout local sync path
-          checkoutFS = scopedFSes.get(archive.localSyncSettings.path)
-          checkoutFS.setFilter(p => folderSync.applyDatIgnoreFilter(archive, p))
-          isPreview = true
-        } else {
-          let err = new Error('Preview mode is not enabled for this dat')
-          err.noPreviewMode = true
-          throw err
-        }
-      } else {
-        throw new Error('Invalid version identifier:' + version)
-      }
-    } else {
-      if (seq <= 0) throw new Error('Version too low')
-      if (seq > archive.version) throw new Error('Version too high')
-      checkoutFS = archive.checkout(seq, {metadataStorageCacheSize: 0, contentStorageCacheSize: 0, treeCacheSize: 0})
-      isHistoric = true
-    }
-  }
+  // DAEMON
+  // if (version) {
+  //   let seq = parseInt(version)
+  //   if (Number.isNaN(seq)) {
+  //     if (version === 'latest') {
+  //       // ignore, we use latest by default
+  //     } else if (version === 'preview') {
+  //       if (archive.localSyncSettings) {
+  //         // checkout local sync path
+  //         checkoutFS = scopedFSes.get(archive.localSyncSettings.path)
+  //         checkoutFS.setFilter(p => folderSync.applyDatIgnoreFilter(archive, p))
+  //         isPreview = true
+  //       } else {
+  //         let err = new Error('Preview mode is not enabled for this dat')
+  //         err.noPreviewMode = true
+  //         throw err
+  //       }
+  //     } else {
+  //       throw new Error('Invalid version identifier:' + version)
+  //     }
+  //   } else {
+  //     if (seq <= 0) throw new Error('Version too low')
+  //     if (seq > archive.version) throw new Error('Version too high')
+  //     checkoutFS = archive.checkout(seq, {metadataStorageCacheSize: 0, contentStorageCacheSize: 0, treeCacheSize: 0})
+  //     isHistoric = true
+  //   }
+  // }
   return {isHistoric, isPreview, checkoutFS}
 }
 
@@ -518,27 +405,9 @@ const getOrLoadArchive = exports.getOrLoadArchive = async function getOrLoadArch
 
 exports.unloadArchive = async function unloadArchive (key) {
   key = fromURLToKey(key)
-  const archive = archives[key]
-  if (!archive) {
-    return
-  }
-
-  // shutdown archive
-  leaveSwarm(key)
-  stopAutodownload(archive)
-  if (archive.fileActStream) {
-    archive.fileActStream.end()
-    archive.fileActStream = null
-  }
-  datExtensions.detach(archive)
-  await new Promise((resolve, reject) => {
-    archive.close(err => {
-      if (err) reject(err)
-      else resolve()
-    })
-  })
-  delete archivesByDKey[datEncoding.toStr(archive.discoveryKey)]
+  if (!archives[key]) return
   delete archives[key]
+  await daemon.unloadArchive(key)
 }
 
 const isArchiveLoaded = exports.isArchiveLoaded = function isArchiveLoaded (key) {
@@ -547,8 +416,9 @@ const isArchiveLoaded = exports.isArchiveLoaded = function isArchiveLoaded (key)
 }
 
 const updateSizeTracking = exports.updateSizeTracking = async function updateSizeTracking (archive) {
+  // DAEMON
   // fetch size
-  archive.size = await pda.readSize(archive, '/')
+  // archive.size = await pda.readSize(archive, '/')
 }
 
 // archive fetch/query
@@ -567,9 +437,9 @@ exports.queryArchives = async function queryArchives (query) {
     var archive = getArchive(archiveInfo.key)
     if (archive) {
       archiveInfo.isSwarmed = archiveInfo.userSettings.networked
-      archiveInfo.size = archive.size
-      archiveInfo.peers = archive.metadata.peers.length
-      archiveInfo.peerHistory = archive.peerHistory
+      archiveInfo.size = 0 // archive.size DAEMON
+      archiveInfo.peers = 0 // archive.metadata.peers.length DAEMON
+      archiveInfo.peerHistory = [] // archive.peerHistory DAEMON
     } else {
       archiveInfo.isSwarmed = false
       archiveInfo.peers = 0
@@ -591,10 +461,10 @@ exports.getArchiveInfo = async function getArchiveInfo (key) {
   ])
   meta.key = key
   meta.url = `dat://${key}`
-  meta.links = archive.manifest.links || {}
-  meta.manifest = archive.manifest
-  meta.version = archive.version
-  meta.size = archive.size
+  meta.links = {} // archive.manifest.links || {} DAEMON
+  meta.manifest = {} // archive.manifest DAEMON
+  meta.version = 1 // archive.version DAEMON
+  meta.size = 1 // archive.size DAEMON
   meta.userSettings = {
     isSaved: userSettings.isSaved,
     hidden: userSettings.hidden,
@@ -605,74 +475,17 @@ exports.getArchiveInfo = async function getArchiveInfo (key) {
     localSyncPath: userSettings.localSyncPath,
     previewMode: userSettings.previewMode
   }
-  meta.peers = archive.metadata.peers.length
-  meta.peerInfo = getArchivePeerInfos(archive)
-  meta.peerHistory = archive.peerHistory
-  meta.networkStats = archive.networkStats
+  meta.peers = 0 // archive.metadata.peers.length DAEMON
+  meta.peerInfo = [] // getArchivePeerInfos(archive) DAEMON
+  meta.peerHistory = [] // archive.peerHistory DAEMON
+  meta.networkStats = {} // archive.networkStats DAEMON
 
   return meta
 }
 
 exports.clearFileCache = async function clearFileCache (key) {
-  var archive = await getOrLoadArchive(key)
-  if (archive.writable) {
-    return // abort, only clear the content cache of downloaded archives
-  }
-
-  // clear the cache
-  await new Promise((resolve, reject) => {
-    archive.content.clear(0, archive.content.length, err => {
-      if (err) reject(err)
-      else resolve()
-    })
-  })
-
-  // force a reconfig of the autodownloader
   var userSettings = await archivesDb.getUserSettings(0, key)
-  stopAutodownload(archive)
-  configureAutoDownload(archive, userSettings)
-}
-
-// archive networking
-// =
-
-// set the networking of an archive based on settings
-function configureNetwork (archive, settings) {
-  if (!settings || settings.networked) {
-    joinSwarm(archive)
-  } else {
-    leaveSwarm(archive)
-  }
-}
-
-// put the archive into the network, for upload and download
-const joinSwarm = exports.joinSwarm = function joinSwarm (key, opts) {
-  var archive = (typeof key === 'object' && key.key) ? key : getArchive(key)
-  if (!archive || archive.isSwarming) return
-  archiveSwarm.join(archive.discoveryKey)
-  var keyStr = datEncoding.toStr(archive.key)
-  log(keyStr, {
-    event: 'swarming',
-    discoveryKey: datEncoding.toStr(archive.discoveryKey)
-  })
-  archive.isSwarming = true
-}
-
-// take the archive out of the network
-const leaveSwarm = exports.leaveSwarm = function leaveSwarm (key) {
-  var archive = (typeof key === 'object' && key.discoveryKey) ? key : getArchive(key)
-  if (!archive || !archive.isSwarming) return
-
-  var keyStr = datEncoding.toStr(archive.key)
-  log(keyStr, {
-    event: 'unswarming',
-    message: `Disconnected ${archive.metadata.peers.length} peers`
-  })
-
-  archive.replicationStreams.forEach(stream => stream.destroy()) // stop all active replications
-  archive.replicationStreams.length = 0
-  archiveSwarm.leave(archive.discoveryKey)
-  archive.isSwarming = false
+  return daemon.clearFileCache(key, userSettings)
 }
 
 // helpers
@@ -711,199 +524,65 @@ const fromKeyToURL = exports.fromKeyToURL = function fromKeyToURL (key) {
   return key
 }
 
-const getLocalSyncSettings = exports.getLocalSyncSettings = function getLocalSyncSettings (archive, userSettings) {
-  if (!archive.writable || !userSettings.isSaved) {
-    return false
-  }
-  if (userSettings.localSyncPath) {
-    return {
-      path: userSettings.localSyncPath,
-      autoPublish: !userSettings.previewMode
-    }
-  }
-  if (userSettings.previewMode) {
-    return {
-      path: archivesDb.getInternalLocalSyncPath(archive),
-      autoPublish: false,
-      isUsingInternal: true
-    }
-  }
-  return false
-}
-
-// internal methods
+// archive proxy
 // =
 
-function configureAutoDownload (archive, userSettings) {
-  if (archive.writable) {
-    return // abort, only used for unwritable
-  }
-  // HACK
-  // mafintosh is planning to put APIs for this inside of hyperdrive
-  // till then, we'll do our own inefficient downloader
-  // -prf
-  const isAutoDownloading = userSettings.isSaved && userSettings.autoDownload
-  if (!archive._autodownloader && isAutoDownloading) {
-    // setup the autodownload
-    archive._autodownloader = {
-      undownloadAll: () => {
-        if (archive.content) {
-          archive.content._selections.forEach(range => archive.content.undownload(range))
+function makeArchiveProxyCbFn (key, method) {
+  return (...args) => daemon.callArchiveAsyncMethod(key, method, ...args)
+}
+
+function makeArchiveProxyReadStreamFn (key, method) {
+  return (...args) => daemon.callArchiveReadStreamMethod(key, method, ...args)
+}
+
+function makeArchiveProxyWriteStreamFn (key, method) {
+  return (...args) => daemon.callArchiveWriteStreamMethod(key, method, ...args)
+}
+
+function createArchiveProxy (key, archiveInfo) {
+  key = datEncoding.toStr(key)
+  const stat = makeArchiveProxyCbFn(key, 'stat')
+  return {
+    key: datEncoding.toBuf(key),
+
+    // DAEMON
+    version: 0,
+    writable: false,
+    discoveryKey: null,
+
+    ready: makeArchiveProxyCbFn(key, 'ready'),
+    download: makeArchiveProxyCbFn(key, 'download'),
+    history: makeArchiveProxyReadStreamFn(key, 'history'),
+    createReadStream: makeArchiveProxyReadStreamFn(key, 'createReadStream'),
+    readFile: makeArchiveProxyCbFn(key, 'readFile'),
+    createDiffStream: makeArchiveProxyReadStreamFn(key, 'createDiffStream'),
+    createWriteStream: makeArchiveProxyWriteStreamFn(key, 'createWriteStream'),
+    writeFile: makeArchiveProxyCbFn(key, 'writeFile'),
+    unlink: makeArchiveProxyCbFn(key, 'unlink'),
+    mkdir: makeArchiveProxyCbFn(key, 'mkdir'),
+    rmdir: makeArchiveProxyCbFn(key, 'rmdir'),
+    readdir: makeArchiveProxyCbFn(key, 'readdir'),
+    stat: (...args) => {
+      var cb = args.pop()
+      args.push((err, st) => {
+        if (st) {
+          // turn into proper stat object
+          st.atime = new Date(st.atime)
+          st.mtime = new Date(st.mtime)
+          st.ctime = new Date(st.ctime)
+          st.isSocket = () => false
+          st.isSymbolicLink = () => false
+          st.isFile = () => st.mode & 32768 === 32768
+          st.isBlockDevice = () => false
+          st.isDirectory = () => st.mode & 16384 === 16384
+          st.isCharacterDevice = () => false
+          st.isFIFO = () => false
         }
-      },
-      onUpdate: throttle(() => {
-        // cancel ALL previous, then prioritize ALL current
-        archive._autodownloader.undownloadAll()
-        pda.download(archive, '/').catch(e => { /* ignore cancels */ })
-      }, 5e3)
-    }
-    archive.metadata.on('download', archive._autodownloader.onUpdate)
-    pda.download(archive, '/').catch(e => { /* ignore cancels */ })
-  } else if (archive._autodownloader && !isAutoDownloading) {
-    stopAutodownload(archive)
-  }
-}
-
-function configureLocalSync (archive, userSettings) {
-  var oldLocalSyncSettings = archive.localSyncSettings
-  archive.localSyncSettings = getLocalSyncSettings(archive, userSettings)
-
-  if (!isEqual(archive.localSyncSettings, oldLocalSyncSettings)) {
-    // configure the local folder watcher if a change occurred
-    folderSync.configureFolderToArchiveWatcher(archive)
-  }
-
-  if (!archive.localSyncSettings || !archive.localSyncSettings.isUsingInternal) {
-    // clear the internal directory if it's not in use
-    jetpack.removeAsync(archivesDb.getInternalLocalSyncPath(archive))
-  }
-}
-
-function stopAutodownload (archive) {
-  if (archive._autodownloader) {
-    archive._autodownloader.undownloadAll()
-    archive.metadata.removeListener('download', archive._autodownloader.onUpdate)
-    archive._autodownloader = null
-  }
-}
-
-function connectReplicationStream (local, remote) {
-  var streams = [local, remote, local]
-  if (upThrottleGroup) streams.splice(1, 0, upThrottleGroup.throttle())
-  if (downThrottleGroup) streams.splice(-1, 0, downThrottleGroup.throttle())
-  pump(streams)
-}
-
-function createReplicationStream (info) {
-  // create the protocol stream
-  var streamKeys = [] // list of keys replicated over the streamd
-  var stream = hypercoreProtocol({
-    id: networkId,
-    live: true,
-    encrypt: true,
-    extensions: ['ephemeral', 'session-data']
-  })
-  stream.peerInfo = info
-
-  // add the archive if the discovery network gave us any info
-  if (info.channel) {
-    add(info.channel)
-  }
-
-  // add any requested archives
-  stream.on('feed', add)
-
-  function add (dkey) {
-    // lookup the archive
-    var dkeyStr = datEncoding.toStr(dkey)
-    var archive = archivesByDKey[dkeyStr]
-    if (!archive || !archive.isSwarming) {
-      return
-    }
-    if (archive.replicationStreams.indexOf(stream) !== -1) {
-      return // already replicating
-    }
-
-    // create the replication stream
-    archive.replicate({stream, live: true})
-    if (stream.destroyed) return // in case the stream was destroyed during setup
-
-    // track the stream
-    var keyStr = datEncoding.toStr(archive.key)
-    streamKeys.push(keyStr)
-    archive.replicationStreams.push(stream)
-    function onend () {
-      archive.replicationStreams = archive.replicationStreams.filter(s => (s !== stream))
-    }
-    stream.once('error', onend)
-    stream.once('end', onend)
-    stream.once('finish', onend)
-    stream.once('close', onend)
-  }
-
-  // debugging
-  stream.on('error', err => {
-    log(streamKeys, {
-      event: 'connection-error',
-      peer: `${info.host}:${info.port}`,
-      connectionType: info.type,
-      message: err.toString()
-    })
-  })
-
-  return stream
-}
-
-function onNetworkChanged (archive) {
-  var now = Date.now()
-  var lastHistory = archive.peerHistory.slice(-1)[0]
-  if (lastHistory && (now - lastHistory.ts) < 10e3) {
-    // if the last datapoint was < 10s ago, just update it
-    lastHistory.peers = archive.metadata.peers.length
-  } else {
-    archive.peerHistory.push({
-      ts: Date.now(),
-      peers: archive.metadata.peers.length
-    })
-  }
-
-  // keep peerHistory from getting too long
-  if (archive.peerHistory.length >= 500) {
-    // downsize to 360 points, which at 10s intervals covers one hour
-    archive.peerHistory = archive.peerHistory.slice(archive.peerHistory.length - 360)
-  }
-
-  // count # of peers
-  var totalPeerCount = 0
-  for (var k in archives) {
-    totalPeerCount += archives[k].metadata.peers.length
-  }
-  archivesEvents.emit('network-changed', {
-    details: {
-      url: `dat://${datEncoding.toStr(archive.key)}`,
-      peers: getArchivePeerInfos(archive),
-      connections: archive.metadata.peers.length,
-      totalPeerCount
-    }
-  })
-}
-
-function getArchivePeerInfos (archive) {
-  // old way, more accurate?
-  // archive.replicationStreams.map(s => ({host: s.peerInfo.host, port: s.peerInfo.port}))
-
-  return archive.metadata.peers.map(peer => peer.stream.stream.peerInfo).filter(Boolean)
-}
-
-function log (key, data) {
-  var keys = Array.isArray(key) ? key : [key]
-  debug(Object.keys(data).reduce((str, key) => str + `${key}=${data[key]} `, '') + `key=${keys.join(',')}`)
-  keys.forEach(k => {
-    let data2 = Object.assign(data, {archiveKey: k})
-    debugEvents.emit(k, data2)
-    debugEvents.emit('all', data2)
-  })
-  if (keys[0]) {
-    debugLogFile.append(keys[0] + JSON.stringify(data) + '\n')
+        cb(err, st)
+      })
+      stat(...args)
+    },
+    lstat: makeArchiveProxyCbFn(key, 'lstat'),
+    access: makeArchiveProxyCbFn(key, 'access')
   }
 }
